Создаёт модульную структу скриптов
